move list view behind toggle in own components
