+++ src/components/InfoOmDigitalSykmelding/InfoOmDigitalSykmelding.tsx
@@ -1,9 +1,9 @@
 import { ReactElement } from 'react'
-import { Accordion, BodyLong, Heading, Link } from '@navikt/ds-react'
+import { Accordion, BodyLong, BodyShort, Heading, Link, LinkPanel } from '@navikt/ds-react'
 
 import { browserEnv } from '../../utils/env'
 
-const InfoOmDigitalSykmelding = (): ReactElement => (
+export const InfoOmDigitalSykmelding = (): ReactElement => (
     <Accordion.Item>
         <Accordion.Header>Om den digitale sykmeldingen</Accordion.Header>
         <Accordion.Content>
@@ -47,4 +47,42 @@ const InfoOmDigitalSykmelding = (): ReactElement => (
     </Accordion.Item>
 )
 
-export default InfoOmDigitalSykmelding
+export function SerIkkeSykmelding(): ReactElement {
+    return (
+        <Accordion.Item>
+            <Accordion.Header>Ser du ikke sykmeldingen din her?</Accordion.Header>
+            <Accordion.Content>
+                <LinkPanel
+                    href="https://person.nav.no/mine-saker/tema/SYM"
+                    target="_blank"
+                    border
+                    className="mb-8 mt-4 rounded-large"
+                >
+                    <LinkPanel.Title className="text-heading-xsmall">Sjekk dokumentlisten</LinkPanel.Title>
+                </LinkPanel>
+
+                <div className="mb-4">
+                    <BodyShort>
+                        Det kan også bety at den som har sykmeldt deg ikke sender den digitalt til NAV. Da bruker du{' '}
+                        <Link
+                            href="https://www.helsedirektoratet.no/veiledere/sykmelderveileder/sykmelding-og-erklaeringer"
+                            target="_blank"
+                        >
+                            papirsykmeldingen
+                        </Link>{' '}
+                        i stedet.
+                    </BodyShort>
+                </div>
+
+                <div className="mb-4">
+                    <BodyShort>
+                        <Link href="https://www.nav.no/kontaktoss" target="_blank">
+                            Kontakt oss
+                        </Link>{' '}
+                        om du fortsatt ikke finner det du leter etter.
+                    </BodyShort>
+                </div>
+            </Accordion.Content>
+        </Accordion.Item>
+    )
+}
