+++ src/components/SykmeldingerList/SykmeldingerListDynamic.tsx
@@ -1,0 +1,7 @@
+import React, { ReactElement } from 'react'
+
+function SykmeldingerListDynamic(): ReactElement {
+    return <div></div>
+}
+
+export default SykmeldingerListDynamic
